Refatorando organização de certos componentes.
Autenticação do usuário e adição de rotas privadas.

+++ src/models/equipment.js
@@ -1,5 +1,5 @@
-const db = require('./db');
-const helper = require('../helper');
+const db = require('../services/db');
+const helper = require('../utils/helper');
 
 async function getAll(page = 1) {
   const offset = helper.getOffset(page, 10);
@@ -14,62 +14,75 @@ async function getAll(page = 1) {
 }
 
 async function add({ nome, unidades }) {
+  if (!(nome && unidades)) {
+    return { message: 'Bad Request, malformed syntax', statusCode: 400 };
+  }
+
   const result = await db.query(
     `INSERT INTO Equipamentos (nome, unidades) 
     values(?,?);`,
     [nome, unidades]
   );
 
-  let message = 'Error in creating user';
+  const id = result.insertId;
 
   if (result.affectedRows) {
-    message = 'User created successfully';
+    return { id, statusCode: 200 };
   }
-  return { message };
+  return { message: 'Error in creating equipment', statusCode: 500 };
 }
 
 async function get(id) {
+  if (!id) {
+    return { message: 'Bad Request, malformed syntax', statusCode: 400 };
+  }
+
   const result = await db.query(
     `SELECT * FROM Equipamentos 
     WHERE idEquipamento=?;`,
     [id]
   );
-  let message = 'Error in finding user';
+
   if (result[0]) {
-    message = 'User found successfully';
+    return { values: result[0], statusCode: 200 };
   }
 
-  return result[0];
+  return { message: 'Equipment not found', statusCode: 404 };
 }
 
 async function update(id, { nome, unidades }) {
+  if (!(id && nome && unidades)) {
+    return { message: 'Bad Request, malformed syntax', statusCode: 400 };
+  }
+
   const result = await db.query(
     `UPDATE Equipamentos SET nome=?, unidades=?
     WHERE idEquipamento=?;`,
     [nome, unidades, id]
   );
-  let message = 'Error in updating user';
 
   if (result.affectedRows) {
-    message = 'User updated successfully';
+    return { id, statusCode: 200 };
   }
 
-  return { message };
+  return { message: 'Error in updating equipment', statusCode: 404 };
 }
 
 async function remove(id) {
+  if (!id) {
+    return { message: 'Bad Request, malformed syntax', statusCode: 400 };
+  }
+
   const result = await db.query(
     `DELETE FROM Equipamentos WHERE idEquipamento=?`,
     [id]
   );
 
-  let message = 'Error in deleting user';
-
   if (result.affectedRows) {
-    message = 'User deleted successfully';
+    return { id, statusCode: 200 };
   }
 
-  return { message };
+  return { message: 'Error in removing equipment', statusCode: 404 };
 }
 
 module.exports = {
